✨ 이벤트리스너 별도함수분리, 클린코드로 함수 분리
